requirements.txt
requests
python-dotenv
beautifulsoup4
langchain
langchain_openai
chromadb
openai